adapter: combine cluster sequencer files
Previously, the logic for sequencing cluster related commands were
split up into 3 files.

  - `src/adapter/src/coord/sequencer/inner/cluster.rs`
  - `src/adapter/src/coord/sequencer/cluster.rs`
  - `src/adapter/src/coord/sequencer/alter_set_cluster.rs`

As far as I can tell there is no rhyme or reason for this split, and it
makes it annoying to navigate through the cluster methods. This commit
combines the files into a single file.
