--- src/adapter/src/coord/sequencer/alter_set_cluster.rs
@@ -1,35 +1,0 @@
-// Copyright Materialize, Inc. and contributors. All rights reserved.
-//
-// Use of this software is governed by the Business Source License
-// included in the LICENSE file.
-//
-// As of the Change Date specified in that file, in accordance with
-// the Business Source License, use of this software will be governed
-// by the Apache License, Version 2.0.
-
-use mz_sql::plan::AlterSetClusterPlan;
-
-use crate::coord::Coordinator;
-use crate::session::Session;
-use crate::{AdapterError, ExecuteResponse};
-
-impl Coordinator {
-    /// Convert a [`AlterSetClusterPlan`] to a sequence of catalog operators and adjust state.
-    pub(super) async fn sequence_alter_set_cluster(
-        &mut self,
-        _session: &Session,
-        AlterSetClusterPlan { id, set_cluster: _ }: AlterSetClusterPlan,
-    ) -> Result<ExecuteResponse, AdapterError> {
-        // TODO: This function needs to be implemented.
-
-        // Satisfy Clippy that this is an async func.
-        async {}.await;
-        let entry = self.catalog().get_entry(&id);
-        match entry.item().typ() {
-            _ => {
-                // Unexpected; planner permitted unsupported plan.
-                Err(AdapterError::Unsupported("ALTER SET CLUSTER"))
-            }
-        }
-    }
-}
